--- apps/banco/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    # path('', views.index, name='index'),
-]
