+++ apps/cartoes/views.py
@@ -3,23 +3,6 @@ from django.shortcuts import render, redirect
 from .models import Cartao
 from django.contrib.auth.decorators import login_required    
 from django.http import JsonResponse 
-from django.core.mail import send_mail
-
-def send_confirmation_email(request):
-    if request.method == 'POST':
-        recipient_email = request.POST.get('email')
-
-        send_mail(
-            subject='Confirmação de Cadastro',
-            message='Seu cadastro foi realizado com sucesso!',
-            from_email=None,  # usa o DEFAULT_FROM_EMAIL
-            recipient_list=[recipient_email],
-            fail_silently=False,
-        )
-
-        return render(request, 'email_sent.html')
-
-    return render(request, 'send_email_form.html')
 
 
 def cartoes(request):
