+++ apps/emails/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path('send_test_email/', views.send_confirmation_email, name='send_test_email'),
+]
